requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-pptx
jinja2
python-dotenv
pytest
httpx
